refactor: changed the Database and Index interface.
To support CotnentTypes adding metadata to the index, the
database interface no longer has index specific methods. Instead,
writing of index data has been broken up into:

An Indexer interface which is responsible for indexing common data,
such as metadata.

An EntryIndexer which is used by stores to add each blob as it comes
in.

This allows functionality to be split in a more meaningful way,
and allows the content type to index metadata as required.