feat: mutate metadata via the http api
`/upload/meta` is the new form of the upload api *(`/upload` soon to
follow)*, and allows you to upload metadata based off of a previous
metadata.

This is done either by specifying the anchor on which an index query
will be run to find the previousMeta, or by manually specifying the
previousMeta and no index query is needed.

This commit also introduces the ContentType field of the Meta struct,
which will allow the upload handler to automatically determine the
ContentType responsible for loading and writing said metadata.

Both querying and automatic content assertion can reduce repeated
write performance due to loading/querying data, but both can be
avoided by manually specifying the required information.