refactor: Changed Meta struct to Json struct, including IndexedFields
Meta was implying that the document was metadata about a blob, however
much of the time this was not going to be true. Plain text with no blob
was being stored and indexed just as much, if not moreso, than blobs
with metadata.

As such, Meta has been renamed to Json.

Along side this change, a JsonMeta struct was added. This struct
serves to provide metadata *about* the Json struct. Currently this
only serves to provide the indexer with information about which fields
are to be indexed, and options within them.

Note that it's important to store this index instruction set as
re-indexing needs to have that information available to it.