refactor: moved non-user supplied structs to store.go
Eg, for easy browsing of files, fixity.go contains what users
most commonly interact with. store.go contains lower level
implementation details about data structure.

With the exception of Version, which was moved to Store.go,
as that is returned on Reads.. so it's still a bit user
facing. /shrug