feat: removed hash function prefix
After much debate, i've decided that future-proofing the hashing
function within the address of the content is not required, and
actually hurts the design if the hash function is changed
without migration.

This decision was mainly based around the distributed goals of
fixity. If the hashing function is configurable, then one node
could have it's function changed while offline and offer no
feasible way to inform the other nodes of the hash function change.

This would lead to nodes storing the same data with different hashes,
duplicating it.

Instead, nodes are going to have to communicate their hashing
function as part of their configuration, and only accept identical
configurations.

To deal with the common future-proofing problem of needing to
change the hashing function, a Store / Node migration needs to
be performed.