Revert "feat: Store now returns whether the content addr was created"
This reverts commit 38b5505bef04f4bb5363cd545054838528cc47db.

I'm reverting this commit because it's the wrong way to achieve
the desired end goal of writing a new block only if the content
is new.

Instead, we need to verify that the blockchain has the blob in it
or not. As that is the single source of truth for the fixity store.