feat: added FindContent() to the Blockchain interface.
A breaking rewrite of how Previous Content is navigated must be
performed. The Blockchain must be the authority on whether or not
Content is accessible, which means navigating backwards must be
done from the Blockchain.

In the near future, an optimized "mutable" blockchain will be based
off of the immutable blockchain to allow for direct linking between
content with proper mutation in the event that content versions
were deleted, etc.